# Optimize coalesceSingle