… test/fixtures/coalesce-bench-single-3848571113.json